+++ breeze-core/src/main/kotlin/domain/text/Locales.kt
@@ -84,24 +84,30 @@
  * Breeze is blowing ...
  **********************************************************************************************************************/
 
-package com.windea.breezeframework.core.constants.text
+package com.windea.breezeframework.core.domain.text
 
-/**Sets of common-used strings.*/
-object Strings {
-	const val empty: String = ""
-	const val space: String = " "
-	const val space2: String = "  "
-	const val space4: String = "    "
+/**Sets of common-used locales.*/
+object Locales {
+	const val english: String = "en"
+	const val french: String = "fr"
+	const val german: String = "de"
+	const val italian: String = "it"
+	const val japanese: String = "ja"
+	const val korean: String = "ko"
+	const val chinese: String = "zh"
+	const val simpleChinese: String = "zh_CN"
+	const val traditionalChinese: String = "zh_TW"
 
-	const val singleQuote: String = "'"
-	const val doubleQuote: String = "\""
-	const val backQuote: String = "`"
-
-	const val separator: String = "/"
-	const val backSeparator: String = "\\"
-
-	const val tab: String = "\t"
-	const val cr: String = "\r"
-	const val lf: String = "\n"
-	const val crlf: String = "\r\n"
+	const val france: String = "fr_FR"
+	const val germany: String = "de_DE"
+	const val italy: String = "it_IT"
+	const val japan: String = "ja_JP"
+	const val korea: String = "ko_KR"
+	const val china: String = "zh_CN"
+	const val prc: String = "zh_CN"
+	const val taiwan: String = "zh_TW"
+	const val uk: String = "en_GB"
+	const val us: String = "en_US"
+	const val canada: String = "en_CA"
+	const val canadaFrench: String = "fr_CA"
 }