+++ breeze-core/src/main/kotlin/domain/text/ReferenceCase.kt
@@ -154,10 +154,10 @@ enum class ReferenceCase(
 	 * * `"$.Category.[0].Name"`
 	 */
 	JsonReference(
-		{ it.removePrefix("$").split('.').dropEmpty().map { s -> s.removeSurrounding("[", "]") } },
-		{ it.removePrefix("$").splitToSequence('.').dropEmpty().map { s -> s.removeSurrounding("[", "]") } },
-		{ it.joinToString(".", "$").wrapIndex() },
-		{ it.joinToString(".", "$").wrapIndex() },
+		{ it.removePrefix("$.").split('.').dropEmpty().map { s -> s.removeSurrounding("[", "]") } },
+		{ it.removePrefix("$.").splitToSequence('.').dropEmpty().map { s -> s.removeSurrounding("[", "]") } },
+		{ it.joinToString(".", "$.").wrapIndex() },
+		{ it.joinToString(".", "$.").wrapIndex() },
 		"""\$.*""".toRegex() //不进行严格验证
 	),