+++ breeze-serializer/src/main/kotlin/SerializerExtensions.kt
@@ -97,19 +97,17 @@ import java.io.*
 //如果直接调用typeOf<T>()，会导致NotImplementedError，因此需要调用使用javaTypeOf<T>()
 
 /**从指定字符串读取指定类型的数据。*/
-@OptIn(ImplicitReflectionSerializer::class)
 inline fun <reified T : Any> Serializer.read(string: String): T {
 	return when(this) {
-		is KotlinSerializer<*> -> delegate.parse(string)
+		is KotlinSerializer<*> -> delegate.decodeFromString(string)
 		else -> read(string, javaTypeOf<T>())
 	}
 }
 
 /**从指定文件读取指定类型的数据。*/
-@OptIn(ImplicitReflectionSerializer::class)
 inline fun <reified T : Any> Serializer.read(file: File): T {
 	return when(this) {
-		is KotlinSerializer<*> -> delegate.parse(file.readText())
+		is KotlinSerializer<*> -> delegate.decodeFromString(file.readText())
 		else -> read(file, javaTypeOf<T>())
 	}
 }