+++ breeze-serializer/src/main/kotlin/impl/json/JsonSerializer.kt
@@ -125,9 +125,8 @@ interface JsonSerializer : Serializer {
 
 		/**配置KotlinxSerializationJson的序列化器。注意需要在使用前配置，并且仅当对应的序列化器适用时才应调用。*/
 		@OptionalApi
-		@OptIn(UnstableDefault::class)
-		fun configureKotlinJson(block: (JsonBuilder) -> Unit) {
-			block(KotlinJsonSerializer.jsonBuilder)
+		fun configureKotlinJson(block: (Json) -> Unit) {
+			block(KotlinJsonSerializer.json)
 		}
 
 		/**配置JacksonJson的序列化器。注意需要在使用前配置，并且仅当对应的序列化器适用时才应调用。*/