+++ breeze-unstable/src/main/kotlin/core/domain/Config.kt
@@ -94,5 +94,5 @@ import com.windea.breezeframework.core.annotations.*
  * 这个接口的实现应当是数据类，且对应的应用类应当至少有两个构造方法。
  * 主要构造方法以其为默认参数，而次要构造方法以其代码块为参数。
  */
-@UnstableImplementationApi
+@UnstableApi
 interface Config