+++ breeze-unstable/src/main/kotlin/core/domain/TEntity.kt
@@ -90,7 +90,7 @@ import com.windea.breezeframework.core.annotations.*
 import java.io.*
 
 /**拥有唯一标识符的泛型实体类。此类的子类应当是开放的。*/
-@UnstableImplementationApi
+@UnstableApi
 @Suppress("METHOD_OF_ANY_IMPLEMENTED_IN_INTERFACE")
 abstract class TEntity<ID> : Serializable {
 	open var id: ID? = null